Added changes to allow calcos to run under Python 3 and numpy 1.10.  Requires Astropy 1.1.1. There are still issues due to the change of the Python round() function from 2 to 3, these will be fixed in a future version
git-svn-id: http://svn.stsci.edu/svn/ssb/stsci_python/calcos/branches/opus_2015_3dms@48906 fe389314-cf27-0410-b35b-8c050e845b92

+++ lib/calcos/concurrent.py
@@ -1330,7 +1330,7 @@ class ConcurrentWavecal(object):
             xi = self.xi_corr[shift_flags]      # copy out the relevant subset
             if len(xi) > 0:
                 xi_diff = xi - np.around(xi)
-                dpixel1 = xi_diff.mean()
+                dpixel1 = xi_diff.mean(dtype=np.float64)
                 value = round(dpixel1, 4)
             else:
                 value = 0.
